pull ApproximateRate out into its own module
